python/requirements.txt
azure-data-tables>=12.4.3
azure-identity>=1.15.0
python-dotenv>=1.0.0
